+++ CHANGES.md
@@ -4,6 +4,8 @@ This file is for tracking changes.
 
 ### Unreleased
 
+New :
+- Add initial database schema
 
 ### 0.1.0 - 2025-11-09
 - Initial release